modify dao and dto

+++ src/main/java/com/groupthree/culinarycompanion/dao/IRecipeDAO.java
@@ -1,10 +1,10 @@
 package com.groupthree.culinarycompanion.dao;
 
 public interface IRecipeDAO {
-    Recipe findRecipeById(long recipeId);
+    Recipe findRecipeById(int recipeId);
     Recipe findRecipeByName(String name);
     void createRecipe(Recipe recipe);
     void updateRecipe(Recipe recipe);
-    void deleteRecipe(long recipeId);
+    void deleteRecipe(int recipeId);
 }
