+++ src/main/java/com/groupthree/culinarycompanion/dao/IUserDAO.java
@@ -1,9 +1,9 @@
 package com.groupthree.culinarycompanion.dao;
 
 public interface IUserDAO {
-    User findUserById(long userId);
+    User findUserById(int userId);
     User findUserByEmail(String email);
     void createUser(User user);
     void updateUser(User user);
-    void deleteUser(long userId);
+    void deleteUser(int userId);
 }
